+++ examples/build (high-performance).py
@@ -1,11 +1,11 @@
 import fiberhttp
 
-cn = fiberhttp.client()
-build = fiberhttp.build('POST', 'httpbin.org', '/post', headers={}, data='number=10')
+cn = fiberhttp.Client()
+build = fiberhttp.Request('POST', 'httpbin.org', '/post', headers={}, data='number=10')
 
 for _ in range(10):
     data = 'number=10'
-    RES = cn.send('httpbin.org', build)
+    RES = cn.send(build)
     
     if RES.text().__contains__(data):
         print('OK')