+++ examples/build proxy (high-performance).py
@@ -1,7 +1,7 @@
 import fiberhttp
 
-cn = fiberhttp.client_proxy('http://Heisenberg:505152@20.111.54.16:8123', proxy_ssl=False)
-build = fiberhttp.build_proxy('POST', 'httpbin.org', '/post', headers={}, data='number=10', proxy_auth=cn.proxy_auth)
+cn = fiberhttp.Client_Proxy('http://Heisenberg:505152@20.111.54.16:8123', proxy_ssl=False)
+build = fiberhttp.Request('POST', 'httpbin.org', '/post', headers={}, data='number=10')
 
 for _ in range(10):
     data = 'number=' + str(_)