+++ examples/json_arg.py
@@ -1,6 +1,6 @@
 import fiberhttp
 
-cn = fiberhttp.client()
+cn = fiberhttp.Client()
 
 response = cn.post('https://httpbin.org/post', json={'username':'ndoshy'}).json()['json']['username'] == 'ndoshy'