+++ fiberhttp/__init__.py
@@ -1,6 +1,6 @@
-from ._client_file import client
-from ._client_proxy import client_proxy
+from ._client_file import Client
+from ._client_proxy import Client_Proxy
 from ._methods import get, post, put, delete, patch
-from ._build import request
+from ._build import Request
 
-__all__ = ['client', 'client_proxy', 'request', 'get', 'post', 'put', 'delete', 'patch']
+__all__ = ['Client', 'Client_Proxy', 'Request', 'get', 'post', 'put', 'delete', 'patch']