+++ fiberhttp/_connections.py
@@ -1,21 +1,38 @@
+from ._exceptions import TimeoutException, ConnectionErrorException
 from typing import Union
 from ssl import SSLSocket, SSLContext, PROTOCOL_TLS_CLIENT
 from socket import socket, AF_INET, SOCK_STREAM
 from certifi import where
 
 def new_connection(host:str, port:int) -> Union[socket, SSLSocket]:
-    connection = socket(AF_INET, SOCK_STREAM)
-    connection.connect((host, port))
+    try:
+        connection = socket(AF_INET, SOCK_STREAM)
+        connection.connect((host, port))
 
-    if port == 443:
-        context : SSLContext = SSLContext(PROTOCOL_TLS_CLIENT)
-        context.load_verify_locations(where())
-        connection = context.wrap_socket(connection, server_hostname=host)
+        if port == 443:
+            context : SSLContext = SSLContext(PROTOCOL_TLS_CLIENT)
+            context.load_verify_locations(where())
+            connection = context.wrap_socket(connection, server_hostname=host)
+
+    except TimeoutError:
+        raise TimeoutException()
+
+    except:
+        ConnectionErrorException()
+    
     return connection
     
 def new_connection_proxy(host:str, port:int) -> socket:
-    connection = socket(AF_INET, SOCK_STREAM)
-    connection.connect((host, port))
+    try:
+        connection = socket(AF_INET, SOCK_STREAM)
+        connection.connect((host, port))
+
+    except TimeoutError:
+        raise TimeoutException()
+    
+    except:
+        ConnectionErrorException()   
+
     return connection
     
 def load_ssl(connection:socket) -> SSLSocket: